+++ next.config.mjs
@@ -1,4 +1,5 @@
 import createMDX from '@next/mdx'
+import { withSentryConfig } from '@sentry/nextjs'
 import remarkGfm from 'remark-gfm'
 import rehypeSlug from 'rehype-slug'
 import rehypePrettyCode from 'rehype-pretty-code'
@@ -43,4 +44,18 @@ const withMDX = createMDX({
   },
 })
 
-export default withBundleAnalyzer(withMDX(nextConfig))
+const sentryWebpackPluginOptions = {
+  silent: true,
+  org: process.env.SENTRY_ORG,
+  project: process.env.SENTRY_PROJECT,
+  authToken: process.env.SENTRY_AUTH_TOKEN,
+  dryRun: !process.env.SENTRY_AUTH_TOKEN,
+}
+
+const compiledConfig = withBundleAnalyzer(withMDX(nextConfig))
+const sentryNextConfig = {
+  hideSourceMaps: true,
+  disableLogger: true,
+}
+
+export default withSentryConfig(compiledConfig, sentryWebpackPluginOptions, sentryNextConfig)